chore(plugin): translator walks are now in separate units
