+++ dbcon/mysql/ha_mcs_execplan_parseinfo_bits.h
@@ -1,0 +1,28 @@
+/* Copyright (C) 2025 MariaDB Corporation
+
+   This program is free software; you can redistribute it and/or
+   modify it under the terms of the GNU General Public License
+   as published by the Free Software Foundation; version 2 of
+   the License.
+
+   This program is distributed in the hope that it will be useful,
+   but WITHOUT ANY WARRANTY; without even the implied warranty of
+   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+   GNU General Public License for more details.
+
+   You should have received a copy of the GNU General Public License
+   along with this program; if not, write to the Free Software
+   Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston,
+   MA 02110-1301, USA. */
+
+#pragma once
+
+#include <cstdint>
+
+namespace cal_impl_if
+{
+  const uint64_t AGG_BIT = 0x01;
+  const uint64_t SUB_BIT = 0x02;
+  const uint64_t AF_BIT = 0x04;
+  const uint64_t CORRELATED = 0x08;
+}
